+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 PolyhedralRelaxations.jl Change Log
 =========================
 
-### staged
+### Staged 
+- Add base name prefixes for constraints as optional arguments
 
 ### v0.3.4
 - Allow for JuMP 1.0
@@ -13,12 +14,13 @@ PolyhedralRelaxations.jl Change Log
 ### v0.3.3
 - Allow for JuMP 0.23
 
+
 ### v0.3.2
 - Allow for JuMP 0.22
 
 ### v0.3.1
 - Documentation updates
-- Adds base name prefixes as optional argument
+- Add base name prefixes as optional argument
 - Memento version update
 
 ### v0.3.0