+++ CHANGELOG.md
@@ -3,6 +3,7 @@ PolyhedralRelaxations.jl Change Log
 
 ### Staged 
 - Add base name prefixes for constraints as optional arguments
+- Reuse partition variables in one relaxation to another 
 
 ### v0.3.4
 - Allow for JuMP 1.0